Trait matrixcompare::comparators::ElementwiseComparator

pub trait ElementwiseComparator<T> {
    type Error: Display;

    // Required methods
    fn compare(&self, x: &T, y: &T) -> Result<(), Self::Error>;
    fn description(&self) -> String;
}

Trait that describes elementwise comparators for assert_matrix_eq!.

Usually you should not need to interface with this trait directly. It is a part of the documentation only so that the trait bounds for the comparators are made public.

Required Associated Types

type Error: Display

Required Methods

fn compare(&self, x: &T, y: &T) -> Result<(), Self::Error>

Compares two elements.

Returns the error associated with the comparison if it failed.

fn description(&self) -> String

A description of the comparator.